requirements.txt
streamlit>=1.35.0
pandas
numpy
matplotlib
requests
smartapi-python
pyotp
